backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
psycopg2-binary>=2.9
celery[redis]>=5.3
redis>=5.0
gunicorn>=21.2
python-decouple>=3.8
dj-database-url>=2.1
